pyproject.toml
[project]
name = "tutkija"
version = "0.0.1"
description = "Moniagenttinen kirjallisuuskatsaus, Tutkija"
readme = "README.md"
authors = [{ name = "Tutkija-tiimi" }]
requires-python = ">=3.11"
dependencies = [
  "typer>=0.12",
  "httpx>=0.27",
  "pydantic>=2.8",
  "pandas>=2.2",
  "pyarrow>=16.1",
  "rapidfuzz>=3.8",
  "feedparser>=6.0",
  "numpy>=1.26",
  "scikit-learn>=1.4",
  "defusedxml>=0.7",
  "google-generativeai",
  "chromadb", # or faiss-cpu
  "rank-bm25", # or fastbm25
  "tiktoken", # or tokenizers
  "bibtexparser>=1.4.1",
]

[project.scripts]
la = "la_pkg.cli:app"

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["tests"]
markers = [
  "network: tests that require outbound HTTP access",
  "integration: slower, multi-component smoke tests",
  "rag: tests for the RAG pipeline",
]

[project.optional-dependencies]
asreview = ["asreview>=1.5"]
dev = [
  "pytest>=8",
  "pytest-mock>=3.12",
  "respx>=0.21",
  "ruff>=0.5",
  "mypy>=1.10",
]
parse = [
  "lxml>=4.9",
]
